+++ pkg/reconciliation/reconcile_racks.go
@@ -2263,6 +2263,11 @@ func (rc *ReconciliationContext) ReconcileAllRacks() (reconcile.Result, error) {
 
 	endpointData := rc.getCassMetadataEndpoints()
 
+	shouldFQLBeEnabled, serverMajorVersion, recResult := parseFQLFromConfig(rc)
+	if recResult.Completed() {
+		return recResult.Output()
+	}
+
 	if recResult := rc.CheckStatefulSetControllerCaughtUp(); recResult.Completed() {
 		return recResult.Output()
 	}
@@ -2354,6 +2359,10 @@ func (rc *ReconciliationContext) ReconcileAllRacks() (reconcile.Result, error) {
 	}
 
 	if recResult := rc.CheckConditionInitializedAndReady(); recResult.Completed() {
+		return recResult.Output()
+	}
+
+	if recResult := SetFullQueryLogging(rc, shouldFQLBeEnabled, serverMajorVersion); recResult.Completed() {
 		return recResult.Output()
 	}